Find Google Chrome binary in the system

If the chrome binary path has not already been set in HEADLESS_CHROME environment variable, the function will try to find the chrome binary on your system using a some hints.

find_chrome_binary()

Value

A character string. The path the chrome binary that will be used by crrri.

Details

Windows

It will look in the registry for an installed version

macOS,

It will return a hard-coded path of Chrome under /Applications.

Linux,

It will search for chromium-browser and google-chrome from the system's PATH variable.

References

From pagedown R package, licence MIT. Source on Github