Chrome Debugging Protocol HTTP methods

When Chromium/Chrome is started in debugging mode, several HTTP endpoints are available. Using these endpoint, one can retrieve information from Chromium/Chrome or send commands, see https://chromedevtools.github.io/devtools-protocol/#endpoints.

fetch_version(host = "localhost", port = 9222, secure = FALSE)

list_targets(host = "localhost", port = 9222, secure = FALSE)

fetch_protocol(host = "localhost", port = 9222, secure = FALSE,
  local = FALSE)

new_tab(host = "localhost", port = 9222, secure = FALSE,
  url = NULL)

activate_target(host = "localhost", port = 9222, secure = FALSE,
  target_id)

close_target(host = "localhost", port = 9222, secure = FALSE,
  target_id)

inspect_target(host = "localhost", port = 9222, secure = FALSE,
  target_id)

Arguments

host
Character scalar, the host name of the application.

port
The remote debugging port (a numeric or a character scalar).

secure
A logical indicating whether a secure protocol shall be used.

local
Logical scalar, indicating whether the local version of the protocol (embedded in crrri) must be used or the protocol must be fetched remotely.

url
URL to open in a new tab.

target_id
Target (or tab) identifier.

Value

fetch_version(), list_targets(), fetch_protocols() and new_tab() return a list. activate_target() and close_target() returns a logical: TRUE is returned when the command succeeds, FALSE is returned if a wrong target_id is provided.

Examples

# \donttest{
chrome <- Chrome$new()

#> Running google-chrome-stable --no-first-run --headless \
#>   '--user-data-dir=/home/travis/.local/share/r-crrri/chrome-data-dir-mwlefood' \
#>   '--remote-debugging-port=9222' --disable-gpu --no-sandbox

# fetch information about headless Chrome
fetch_version()

#> $Browser
#> [1] "HeadlessChrome/77.0.3865.120"
#> 
#> $`Protocol-Version`
#> [1] "1.3"
#> 
#> $`User-Agent`
#> [1] "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/77.0.3865.120 Safari/537.36"
#> 
#> $`V8-Version`
#> [1] "7.7.299.13"
#> 
#> $`WebKit-Version`
#> [1] "537.36 (@416d6d8013e9adb6dd33b0c12e7614ff403d1a94)"
#> 
#> $webSocketDebuggerUrl
#> [1] "ws://localhost:9222/devtools/browser/0a2c61ba-7935-4d8c-aea4-0bc37a718b32"
#> 

# fetch the Chromium/Chrome protocol
protocol <- fetch_protocol()

# get the list of opened tabs
list_targets()

#> [[1]]
#> [[1]]$description
#> [1] ""
#> 
#> [[1]]$devtoolsFrontendUrl
#> [1] "/devtools/inspector.html?ws=localhost:9222/devtools/page/30286A7DF66FF28EF623B1E7D53C5973"
#> 
#> [[1]]$id
#> [1] "30286A7DF66FF28EF623B1E7D53C5973"
#> 
#> [[1]]$title
#> [1] "about:blank"
#> 
#> [[1]]$type
#> [1] "page"
#> 
#> [[1]]$url
#> [1] "about:blank"
#> 
#> [[1]]$webSocketDebuggerUrl
#> [1] "ws://localhost:9222/devtools/page/30286A7DF66FF28EF623B1E7D53C5973"
#> 
#> 

# open an url in a new tab
new_target <- new_tab(url = "http://r-project.org")
# the new tab is referenced in the list of opened tabs
list_targets()

#> [[1]]
#> [[1]]$description
#> [1] ""
#> 
#> [[1]]$devtoolsFrontendUrl
#> [1] "/devtools/inspector.html?ws=localhost:9222/devtools/page/7895BAC22384D58A826F0CC99094C351"
#> 
#> [[1]]$id
#> [1] "7895BAC22384D58A826F0CC99094C351"
#> 
#> [[1]]$title
#> [1] ""
#> 
#> [[1]]$type
#> [1] "page"
#> 
#> [[1]]$url
#> [1] "http://r-project.org/"
#> 
#> [[1]]$webSocketDebuggerUrl
#> [1] "ws://localhost:9222/devtools/page/7895BAC22384D58A826F0CC99094C351"
#> 
#> 
#> [[2]]
#> [[2]]$description
#> [1] ""
#> 
#> [[2]]$devtoolsFrontendUrl
#> [1] "/devtools/inspector.html?ws=localhost:9222/devtools/page/30286A7DF66FF28EF623B1E7D53C5973"
#> 
#> [[2]]$id
#> [1] "30286A7DF66FF28EF623B1E7D53C5973"
#> 
#> [[2]]$title
#> [1] "about:blank"
#> 
#> [[2]]$type
#> [1] "page"
#> 
#> [[2]]$url
#> [1] "about:blank"
#> 
#> [[2]]$webSocketDebuggerUrl
#> [1] "ws://localhost:9222/devtools/page/30286A7DF66FF28EF623B1E7D53C5973"
#> 
#> 

# inspect a target in a web browser
if(interactive()) inspect_target(target_id = new_target$id)
# close the previous created tab
close_target(target_id = new_target$id)

#> [1] TRUE

chrome$close()
# }