Launch Chromium or Chrome

This class aims to launch Chromium or Chrome in headless mode. It possesses methods to manage connections to headless Chromium/Chrome using the Chrome Debugging Protocol.

Usage

remote <- Chrome$new(bin = Sys.getenv("HEADLESS_CHROME"), debug_port = 9222L,
                     local = FALSE, extra_args = NULL, headless = TRUE,
                     retry_delay = 0.2, max_attempts = 15L)
    remote$connect(callback = NULL)
remote$listConnections()
remote$closeConnections()
remote$version()
remote$user_agent
    remote$close()
remote$view()
remote$is_alive()

Arguments

remote: Chrome object representing a remote instance of headless Chromium/Chrome.

bin: Character scalar, the path to Chromium or Chrome executable.

debug_port: Integer scalar, the Chromium/Chrome remote debugging port. Note that headless Chromium/Chrome will be available at http://localhost:<debug_port>.

local: Logical scalar, indicating whether the local version of the protocol (embedded in crrri) must be used or the protocol must be fetched remotely.

extra_args: Character vector, extra command line arguments passed to Chromium/Chrome.

headless: Logical scalar, indicating whether Chromium/Chrome is launched in headless mode.

retry_delay: Number, delay in seconds between two successive tries to connect to headless Chromium/Chrome.

max_attempts: Logical scalar, number of tries to connect to headless Chromium/Chrome.

callback: Function with one argument, executed when the R session is connected to Chrome. The connection object is passed to this function.

.target_id: A character scalar, identifier of the tab. The default value corresponds to the last created tab. For advanced use only.

Details

$new() opens a new headless Chromium/Chrome.

$connect(callback = NULL, .target_id = "default") connects the R session to the remote instance of headless Chromium/Chrome. The returned value depends on the value of the callback argument. When callback is a function, the returned value is a connection object. When callback is NULL the returned value is a promise which becomes fulfilled once R is connected to the remote instance of Chromium/Chrome. Once fulfilled, the value of this promise is the connection object.

$listConnections() returns a list of the connection objects succesfully created using the $connect() method.

$closeConnections(callback = NULL) closes all the connections created using the $connect() method. Returns a promise which is resolved when all connections are closed.

$version() executes the DevTools Version method. It returns a list of informations available at http://localhost:<debug_port>/json/version.

$user_agent returns a character scalar with the User Agent of the headless Chromium/Chrome.

$close() closes the remote instance of headless Chromium/Chrome.

$view() opens a visible Chromium/Chrome browser at http://localhost:<debug_port>. This is useful to 'see' the headless Chromium/Chrome instance. Returns the process of the visible browser.

$is_alive() checks if the remote instance is alive. Returns a logical scalar.

$listTargets() returns a list with information about tabs.

Examples

# NOT RUN {
remote <- Chrome$new()

remote$connect() %...>% (function(client) {
  Page <- client$Page
  Runtime <- client$Runtime

  Page$enable() %...>% {
    Page$navigate(url = 'http://r-project.org')
  } %...>% {
    Page$loadEventFired()
  } %...>% {
    Runtime$evaluate(
      expression = 'document.documentElement.outerHTML'
    )
  } %...>% (function(result) {
    cat(result$result$value, "\n")
  })
}) %...!% {
  cat("Error:", .$message, "\n")
} %>%
  promises::finally(~ remote$close())
# }